login_register_feature: remove boilerplate and adding a new AuthApi class
