--- api/src/main/java/com/jmanday/api/login/LoginApi.kt
@@ -1,8 +1,0 @@
-package com.jmanday.api.login
-
-import com.jmanday.client.RetrofitClient
-
-class LoginApi {
-
-    val service = RetrofitClient("https://sparky.pl-coding.com:8081/").retrofit.create(LoginService::class.java)
-}
