--- api/src/main/java/com/jmanday/api/login/LoginService.kt
@@ -1,14 +1,0 @@
-package com.jmanday.api.login
-
-import retrofit2.http.Body
-import retrofit2.http.Headers
-import retrofit2.http.POST
-
-interface LoginService {
-    @Headers(
-        "x-api-key: Aj8vfl8S5QTEkhUD"
-    )
-    @POST("login")
-    fun login(@Body loginRequest: LoginRequest): LoginResponseDto
-
-}
