--- api/src/main/java/com/jmanday/api/register/RegisterService.kt
@@ -1,10 +1,0 @@
-package com.jmanday.api.register
-
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface RegisterService {
-
-    @POST("register")
-    suspend fun register(@Body registerRequest: RegisterRequest): RegisterResponseDto
-}
