+++ login/src/main/java/com/jmanday/login/data/datasource/RemoteDataSource.kt
@@ -1,11 +1,11 @@
 package com.jmanday.login.data.datasource
 
-import com.jmanday.api.login.LoginApi
+import com.jmanday.api.AuthApi
 import com.jmanday.api.login.LoginRequest
 
 class RemoteDataSource : UserDataSource {
 
-    private val loginApi = LoginApi()
+    private val loginApi = AuthApi()
 
     override fun login(email: String, password: String) {
         val response = loginApi.service.login(LoginRequest("jmanday@gmail.com", "12a34b56D"))
